requirements.txt
streamlit
openai
python-dotenv
fpdf2
pandas
xlsxwriter
